version en python/requirements.txt
fastapi==0.118.0
uvicorn==0.36.0
sqlalchemy==2.0.43
pydantic==2.11.7
python-jose==3.5.0
passlib==1.7.4
bcrypt==3.2.2
